Completion of multi-segment interface for testing:
wrote code to archive list of output graphs so that the same graph list is available for all COMIDs in the multi-segment system.
Added titles and units to multi-segment graphs and CSV outputs and added option to save CSV to file.
Added warning if the master-setup dates have been changed after linked segments have been created-- similar warning to "choose template"  Added a legend to the flowchart.
Added navigation prompts for the map and the flowchart.
Fixed an issue of rescaling when the label comids button was checked.
Fixed graphical run progress to put it in a threadsafe code.
Improved the flowchart to avoid overlap of branches.
Completed Edits to users manual for multi-segment.
Added error trapping if blank data added to time-series input loading.
Added workaround in case webservice database can't find COMID for map plotting to allow other segments to be plotted.